feat: only show pages that exists in core features
Hide match list page
Hide chat page
Change the announcement list main page from home_page to announcements_page
