+++ BilFoot_flutter/bilfoot/lib/main.dart
@@ -4,7 +4,7 @@ import 'package:bilfoot/views/screens/auth_page/auth_verification_page.dart';
 import 'package:bilfoot/views/screens/auth_page/bloc/authentication_bloc.dart';
 import 'package:bilfoot/views/screens/defining_page/bloc/defining_bloc.dart';
 import 'package:bilfoot/views/screens/defining_page/defining_page.dart';
-import 'package:bilfoot/views/screens/home_page/bloc/announcement_bloc.dart';
+import 'package:bilfoot/views/screens/announcements_page//bloc/announcement_bloc.dart';
 import 'package:bilfoot/views/screens/main_page/main_control_page.dart';
 import 'package:bilfoot/views/screens/match_page/bloc/match_bloc.dart';
 import 'package:bilfoot/views/screens/splash_page/splash_page.dart';
